✨feat: delete hiroshima event area and add history
